refactor: Move the useScannerI18n hook to its Provider file
