+++ src/components/Contexts/SearchProvider.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useScannerI18n } from 'src/components/Hooks/useScannerI18n'
+import { useScannerI18n } from 'src/components/Contexts/ScannerI18nProvider'
 import SearchProvider from 'src/components/Search/SearchProvider'
 
 import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
