+++ src/components/StepperDialog/StepperDialogTitle.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useScannerI18n } from 'src/components/Hooks/useScannerI18n'
+import { useScannerI18n } from 'src/components/Contexts/ScannerI18nProvider'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 
 import Typography from 'cozy-ui/transpiled/react/Typography'
