+++ src/components/Views/PlaceholdersSelector.jsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react'
 import { useNavigate, useLocation, useSearchParams } from 'react-router-dom'
 import { usePapersCreated } from 'src/components/Contexts/PapersCreatedProvider'
 import { usePapersDefinitions } from 'src/components/Contexts/PapersDefinitionsProvider'
-import { useScannerI18n } from 'src/components/Hooks/useScannerI18n'
+import { useScannerI18n } from 'src/components/Contexts/ScannerI18nProvider'
 import FileIcon from 'src/components/Icons/FileIcon'
 import { KonnectorIcon } from 'src/components/Icons/KonnectorIcon'
 import { isReminder } from 'src/components/Placeholders/helpers'
